application/web/requirements.txt
flask-restful==0.3.8
flask_sqlalchemy==2.4.4
mysql-connector-python==8.0.22
flask-serialize==1.4.0